move to right files
